image: Switch Mini-USB port from OTG to device
With OTG and ID detection enabled a process called usb-hardware-sc is
constantly in I/O wait state which results in a load constantly bigger
than 1. This is not a real problem, as that process is not using the CPU.
Linux just includes processes in uninterruptable sleep state into the
load as well.

Also the RED Brick hardware does not support host mode on the Mini-USB
port anyway.