gps-time: Readd support for GPS Bricklet (1.0) and fix date call
Only use GPS datetime if ot had a fix. Don't try to guess if the GPS
datetime is valid.

Pass full GPS datetime resolution to date command.

No need to compensate for time delay between getting GPS datetime and
setting system datetime. The typical delay is in the order of 10ms and
the GPS datetime has a resolution of 100ms.